wibble: return size_t instead of unsigned int like all the other STL containers do

+++ wibble/empty.h
@@ -51,7 +51,7 @@ public:
 	};
 	
 	bool empty() const { return true; }
-	unsigned int size() const { return 0; }
+	size_t size() const { return 0; }
 
 	iterator begin() { return iterator(); }
 	iterator end() { return iterator(); }
